chore: Refactor tests
