--- tests/Testcontainers.Tests/Unit/Configurations/UntilHostTcpPortIsAvailableTest.cs
@@ -1,69 +1,0 @@
-namespace DotNet.Testcontainers.Tests.Unit
-{
-  using System.Threading.Tasks;
-  using DotNet.Testcontainers.Builders;
-  using DotNet.Testcontainers.Commons;
-  using DotNet.Testcontainers.Configurations;
-  using DotNet.Testcontainers.Containers;
-  using JetBrains.Annotations;
-  using Xunit;
-
-  public sealed class UntilExternalTcpPortIsAvailableTest : IAsyncLifetime
-  {
-    private const int listeningPort = 49152;
-    private const int mappedPort = 49153;
-    private const int unmappedPort = 49154;
-    private readonly IContainer _container = new ContainerBuilder()
-      .WithImage(CommonImages.Socat)
-      .WithCommand("-v")
-      .WithCommand($"TCP-LISTEN:{listeningPort},crlf,reuseaddr,fork")
-      .WithCommand("EXEC:cat")
-      .WithPortBinding(listeningPort, true)
-      .WithPortBinding(mappedPort, true)
-      .WithWaitStrategy(Wait.ForUnixContainer().UntilExternalTcpPortIsAvailable(listeningPort))
-      .Build();
-
-    public async ValueTask InitializeAsync()
-    {
-      await _container.StartAsync()
-        .ConfigureAwait(false);
-    }
-
-    public ValueTask DisposeAsync()
-    {
-      return _container.DisposeAsync();
-    }
-
-    [Fact]
-    public async Task SucceededWhenPortIsAvailable()
-    {
-      var untilHostTcpPortIsAvailable = new UntilExternalTcpPortIsAvailable(listeningPort);
-
-      var succeeded = await untilHostTcpPortIsAvailable.UntilAsync(_container);
-
-      Assert.True(succeeded);
-    }
-
-    /// <summary>
-    /// Verifies that the Wait fails when the port is mapped but not listening.
-    /// This test might fail in docker configurations where mapped ports are always listened eg. DockerForMac (https://forums.docker.com/t/port-forwarding-of-exposed-ports-behaves-unexpectedly/15807/2)
-    /// </summary>
-    [Fact(Skip = "The GitHub Action runner allows establishing a connection to any mapped port.")]
-    public async Task FailsWhenPortNotListening()
-    {
-      var untilHostTcpPortIsAvailable = new UntilExternalTcpPortIsAvailable(mappedPort);
-
-      var succeeded = await untilHostTcpPortIsAvailable.UntilAsync(_container);
-
-      Assert.False(succeeded);
-    }
-
-    [Fact]
-    public async Task FailsWhenPortNotMapped()
-    {
-      var untilHostTcpPortIsAvailable = new UntilExternalTcpPortIsAvailable(unmappedPort);
-
-      await Assert.ThrowsAsync<System.InvalidOperationException>(() => untilHostTcpPortIsAvailable.UntilAsync(_container));
-    }
-  }
-}
